+++ log4ts/log4ts.d.ts
@@ -32,6 +32,18 @@ declare module Log4Ts {
     }
 
 
+    interface ConsoleLite {
+        debug: typeof Console.prototype["debug"]; //(message?: any, ...optionalParams: any[]): void;
+        error: typeof Console.prototype["error"]; //(message?: any, ...optionalParams: any[]): void;
+        group: typeof Console.prototype["group"]; //(groupTitle?: string, ...optionalParams: any[]): void;
+        groupEnd: typeof Console.prototype["groupEnd"]; //(): void;
+        info: typeof Console.prototype["info"]; //(message?: any, ...optionalParams: any[]): void;
+        log: typeof Console.prototype["log"]; //(message?: any, ...optionalParams: any[]): void;
+        trace: typeof Console.prototype["trace"]; //(message?: any, ...optionalParams: any[]): void;
+        warn: typeof Console.prototype["warn"]; //(message?: any, ...optionalParams: any[]): void;
+    }
+
+
     export interface EventListener {
         (eventHandler: EventSupport, eventType: string, args: any[] | any): void;
     }
@@ -61,17 +73,7 @@ declare module Log4Ts {
     }
 
 
-    export interface Layout {
-        defaults: {
-            loggerKey: string;
-            timeStampKey: string;
-            millisecondsKey: string;
-            levelKey: string;
-            messageKey: string;
-            exceptionKey: string;
-            urlKey: string;
-        };
-
+    export interface LayoutKeys {
         loggerKey: string;
         timeStampKey: string;
         millisecondsKey: string;
@@ -79,6 +81,12 @@ declare module Log4Ts {
         messageKey: string;
         exceptionKey: string;
         urlKey: string;
+    }
+
+
+    export interface Layout extends LayoutKeys {
+        defaults: LayoutKeys;
+        // extends LayoutKeys
         batchHeader: string;
         batchFooter: string;
         batchSeparator: string;